fixed comoser analyzer test issues

+++ src/Api/ApigeeX/Entity/Property/NanosPropertyAwareTrait.php
@@ -33,7 +33,7 @@ trait NanosPropertyAwareTrait
      */
     public function getNanos(): ?float
     {
-        $nanos = $this->nanos ?? 0;
+        $nanos = (float) ($this->nanos ?? 0);
 
         return $nanos * pow(10, -9);
     }
